Client Credential Flow means client service uses the resource service to protect data
now how things works

So first I'll tell you we have set up two applications.

One is a client application that wishes to access a protected resource from the resource server.

And we have a resource server that we have set up which has a protected resource.

Now what is the protected resource?

It's a very simple endpoint that we have created which is data controller at slash data.

This is authenticated.

How is it authenticated because of this security config.

So we have said like any request it should be authenticated.

And we are making use of OAuth two flow over here.

That is what we have defined over here okay.

And it has marked this particular application as OAuth two resource server as well okay.

Then comes the application properties.

So here we have defined the URL or the configuration from where the token validation for JWT can happen

okay.

So that's the only setting that we need in OAuth two resource server okay.

Coming to the client over here.

So let's talk about the client now in client what we did is we first set up the application over here

like we registered the client ID, secret authorization, grant type, scope, and the URI for tokens.

Okay, so we have added all of this URI for token is needed so that it can get the tokens because it

will need token to pass it over to the resource server so that it can allow the request after validating

if it's a valid request.

Okay.

So these settings we needed over here, then since it's a simple client application which is demonstrating

machine to machine communication, I have not created any separate files.

And within this main application file itself via command line runner.

What we are doing is we are doing an API call and we are fetching the data.

Okay.

We have kept it really simple and to the point.

Now let me explain.

What are we doing.

So we are first exposing the been of rest template over here so that we can make the HTTP client.

Or we can make use of Rest template as a HTTP client for API calls.

Then we have this second method over here.

This one which registers a service that stores authorized clients and tokens in memory.

And it is making use of this inbuilt thing called client registration repository.

It is an interface.

So this is done.

And this is this will actually hold the client that we have defined in the application dot properties

over here.

So that is what is being exposed as been in memory.

And then we have the third method over here which is auth to authorize client manager.

Now what this is going to do is this is going to basically create a manager that will know how

to get access token for registered clients.

So you can see it creates a manager over here.

Okay.

Then it is going to create a provider over here.

Okay.

So it again configures over here a provider is what I can say to use client credentials.

Grant flow.

Okay.

And then it is configuring the manager to make use of the provider.

And then it is returning the manager over here.

Okay.

So so that is what we are doing over here with this particular uh method.

Okay.

Then if you come down over here, we have this command line runner.

Now this will automatically run.

This method will automatically run when the spring boot application starts.

Okay.

There is no API call needed to be done to this application.

Okay.

So you are basically taking, uh, authorized client manager Rest template to make API calls and then,

uh, service URL so that you know where your service is residing.

This is coming in from application dot properties.

Okay.

Now over here what you're doing is you're first creating or crafting the auth request.

So you are creating a OAuth to authorize request using Keycloak Keycloak client as your, uh, client

registration ID and principal, which is the name.

And then you're building this object and then you are getting the client, you're getting the token,

you're creating the headers, and then you are finally making the request over here.

And whatever response you're getting, you are printing that particular response over here.

I hope, uh, this is 100% making sense to you all as to what is happening, why it is happening, and

how things are working in the hindsight.

Now, here, if you take a look at the response that we got, we got this.

We didn't just get the message, we actually got like the status code and everything. Right?

So if you just want to get the message, what you can do is you can say response dot, get body over here.

And now if you run this let me show you the difference.

So now if you run this you will just get the output of the API and not everything else, the status

code and all you can see.

Hello from Service2.

That is all what you have got over here.

Okay, so you can get the body you have access to others.

You can see get body, get headers, get status code.

You have access to all these methods okay.

So now here what is happening is since you are making use of command line runner, this is being executed

on application startup.

But if you are having a service like reporting service you can schedule this, this code, this piece

of code to run on regular intervals like for example, every hour that can be done okay.

And there is no API that needs to be invoked.

It's a job that is being executed every hour, okay?

And it is authenticating itself by itself.

And it is accessing the authenticated APIs on the resource server.

So that's the use case of this, you know, but since we don't want to execute this piece of code

every hour and we don't want to schedule it, we are simply making use of command line runner.